Rename the handlers to have clearer meaning
Summary:
The naming is relative confusing as they have very close names.
It gives me a bit hard time every time I read through the codes.
Therefore, here I sincerely suggest a renaming refactoring.

Test Plan:
1. make test && all green! Hooray!
2. python server.py && check image processing works
